pyproject.toml
[tool.poetry]
name = "webman-analysis"
version = "0.1.0"
description = ""
authors = ["Methuselah-Nwodobeh <methuselah.nwodobeh@amalitech.com>"]
readme = "README.md"
packages = [{include = "webman_analysis"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-decouple = "^2.1"
django-cors-headers = "^4.6.0"
django-prometheus = "^2.3.1"
drf-yasg = "^1.21.8"
psycopg2-binary = "^2.9.10"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
